PATCH: [perl #130921]: Don't use Newx in decoding UTF-8
The bottom level UTF-8 decoding routine can be used during periods when
using Newx is prohibited, as diagnosed by Dave Mitchell for this ticket.
The particular use of Newx was unnecessary, as it is just large enough
to hold a single character, and that can be done by an automatic
variable on the C stack, and probably more efficiently.